FINERACT-1724: Put user notification feature behind feature toggle
