+++ fineract-provider/src/main/java/org/apache/fineract/notification/eventandlistener/NotificationEventListener.java
@@ -18,46 +18,19 @@
  */
 package org.apache.fineract.notification.eventandlistener;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.apache.fineract.notification.data.NotificationData;
-import org.apache.fineract.notification.service.NotificationWritePlatformService;
-import org.apache.fineract.useradministration.domain.AppUser;
-import org.apache.fineract.useradministration.domain.AppUserRepository;
+import org.apache.fineract.notification.service.UserNotificationService;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class NotificationEventListener {
 
-    private final NotificationWritePlatformService notificationWritePlatformService;
-    private final AppUserRepository appUserRepository;
+    private final UserNotificationService userNotificationService;
 
     public void receive(NotificationData notificationData) {
-        Long appUserId = notificationData.getActorId();
+        userNotificationService.notifyUsers(notificationData);
 
-        Set<Long> userIds = notificationData.getUserIds();
-
-        if (notificationData.getOfficeId() != null) {
-            List<Long> tempUserIds = new ArrayList<>(userIds);
-            for (Long userId : tempUserIds) {
-                AppUser appUser = appUserRepository.findById(userId).orElseThrow();
-                if (!Objects.equals(appUser.getOffice().getId(), notificationData.getOfficeId())) {
-                    userIds.remove(userId);
-                }
-            }
-        }
-
-        // Don't notify the same user who triggered the event
-        if (userIds.contains(appUserId)) {
-            userIds.remove(appUserId);
-        }
-
-        notificationWritePlatformService.notify(userIds, notificationData.getObjectType(), notificationData.getObjectId(),
-                notificationData.getAction(), notificationData.getActorId(), notificationData.getContent(),
-                notificationData.isSystemGenerated());
     }
 }
